feat(web): add export page
Remove export button and modal, add export page and navbar link
