+++ packages_rs/nextclade-web/src/components/Main/QuerySequenceList.tsx
@@ -3,10 +3,10 @@ import { Button } from 'reactstrap'
 import styled, { useTheme } from 'styled-components'
 import { ImCross } from 'react-icons/im'
 import { AlgorithmInput } from 'src/types'
-import { ButtonTransparent } from 'src/components/Common/ButtonTransparent'
-import { useTranslationSafe } from 'src/helpers/useTranslationSafe'
 import { useQuerySeqInputs } from 'src/state/inputs.state'
-import { UlGeneric } from '../Common/List'
+import { useTranslationSafe } from 'src/helpers/useTranslationSafe'
+import { ButtonTransparent } from 'src/components/Common/ButtonTransparent'
+import { UlGeneric } from 'src/components/Common/List'
 
 export function QuerySequenceList() {
   const { t } = useTranslationSafe()
