+++ packages_rs/nextclade-web/src/pages/export.tsx
@@ -1,0 +1,1 @@
+export { ExportPage as default } from 'src/components/Export/ExportPage'
